# docs/swagger.yaml
basePath: /
definitions:
  handlers.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  task.Task:
    description: Структура задачи на сервере.
    properties:
      author:
        description: '@Description Автор задачи.'
        type: string
      id:
        description: '@Description Id задачи.'
        type: integer
      text:
        description: '@Description Содержимое задачи.'
        type: string
      title:
        description: '@Description Название задачи.'
        type: string
      urgent:
        description: '@Description Важность задачи(является ли она срочной или нет)'
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for managing tasks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Task API
  version: "1.0"
paths:
  /task:
    get:
      description: Дает содержимое обо всех задачах, название которых начитнается
        с аргумента
      parameters:
      - description: Фильтр по названию
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all tasks by title
      tags:
      - handlers
    post:
      consumes:
      - application/json
      description: Создает задачу по json из тела запроса
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Post task
      tags:
      - handlers
  /task/{id}:
    delete:
      description: Удаляет задачу по ее айди.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete task by ID
      tags:
      - handlers
    get:
      description: Получает задачи по ее айди.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get task by ID
      tags:
      - handlers
    patch:
      description: Обновляет задачу по ее айди, не удаляя старый и создавая новый
        объекты,
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Patch task by ID
      tags:
      - handlers
swagger: "2.0"